pyproject.toml
[project]
name = "pose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deep-sort-realtime>=1.3.2",
    "loguru>=0.7.3",
    "mediapipe>=0.10.31",
    "numpy>=2.4.0",
    "onnx>=1.20.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyside6>=6.10.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.241",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff"
]
